+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/masamerc/sevp/cmd"
+import "github.com/masamerc/sevp/src/cmd"
 
 func main() {
 	cmd.Execute()
