+++ src/cmd/init.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/masamerc/sevp/internal"
+	"github.com/masamerc/sevp/src/internal"
 	"github.com/spf13/cobra"
 )
