+++ src/internal/hooks.go
